# api/swagger/swagger.yml
swagger: "2.0"
info:
  title: API-service
  description: API service for repo and answers
  version: 1.0.0
schemes:
  - https
  - http
host: localhost
basePath: /api
produces:
  - application/json
consumes:
  - application/json
paths:
  /questions/:
    get:
      description: Get a list of repo
      produces:
        - application/json
      responses:
        200:
          description: Questions response
          schema:
            type: array
            items:
              $ref: '#/definitions/Question'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: Internal server
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      summary: Create question
      description: Create a new question
      produces:
        - application/json
      parameters:
        - in: body
          name: question
          required: true
          schema:
            $ref: '#/definitions/QuestionCreate'
      responses:
        201:
          description: Question is created
          schema:
            $ref: '#/definitions/Question'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: Internal server
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  Question:
    type: object
    required:
      - id
      - text
      - created_at
    properties:
      id:
        type: integer
        format: int64
        description: Question id
      text:
        type: string
        description: Question text
      created_at:
        type: string
        format: date-time
        description: Question creation date
  QuestionCreate:
    type: object
    required:
      - text
    properties:
      text:
        type: string
        description: Question text
...
  ErrorResponse:
    type: object
    description: Error type
    properties:
      code:
        type: integer
        description: Error code
      message:
        type: string
        description: Error message
      detail:
        type: string
        description: Error additional details